refactor config a bit
