+++ config/vault_auth.go
@@ -5,8 +5,51 @@ import (
 
 	"github.com/hashicorp/hcl"
 	"github.com/hashicorp/hcl/hcl/ast"
+	"github.com/hashicorp/vault/api"
 	"github.com/mitchellh/mapstructure"
 )
+
+// Auth struct ...
+type Auth struct {
+	Environment     *Environment
+	Name            string
+	Type            string
+	Description     string
+	DefaultLeaseTTL string
+	MaxLeaseTTL     string
+	Config          []*AuthConfig
+	Roles           []*AuthRole
+}
+
+// AuthInput ...
+func (m *Mount) AuthInput() *api.MountInput {
+	return &api.MountInput{
+		Type:        m.Type,
+		Description: m.Description,
+	}
+}
+
+// VaultAuths struct
+//
+// environment
+type VaultAuths []*Auth
+
+// Add ...
+func (m *VaultAuths) Add(auth *Auth) {
+	*m = append(*m, auth)
+}
+
+// AuthConfig ...
+type AuthConfig struct {
+	Name string
+	Data map[string]interface{}
+}
+
+// AuthRole ...
+type AuthRole struct {
+	Name string
+	Data map[string]interface{}
+}
 
 func (c *Config) processVaultAuths(list *ast.ObjectList, environment *Environment) error {
 	if len(list.Items) == 0 {
